chore: export all of ERTP from index.js. Update all imports.
A few other review-suggested cleanups
