+++ packages/ERTP/package.json
@@ -2,7 +2,7 @@
   "name": "@agoric/ertp",
   "version": "0.6.0",
   "description": "Electronic Rights Transfer Protocol (ERTP). A smart contract framework for exchanging electronic rights",
-  "main": "src/issuer.js",
+  "main": "src/index.js",
   "engines": {
     "node": ">=11.0"
   },
