--- packages/ERTP/src/mathHelpersLib.js
@@ -1,20 +1,0 @@
-// @ts-check
-
-import { importManager } from '@agoric/import-manager';
-import natMathHelpers from './mathHelpers/natMathHelpers';
-import strSetMathHelpers from './mathHelpers/strSetMathHelpers';
-import setMathHelpers from './mathHelpers/setMathHelpers';
-
-import './types';
-
-const manager = importManager();
-/**
- * @type {Record<AmountMathKind,MathHelpers>}
- */
-const mathHelpersLib = manager.addExports({
-  nat: natMathHelpers,
-  strSet: strSetMathHelpers,
-  set: setMathHelpers,
-});
-
-export default mathHelpersLib;
