+++ packages/dapp-svelte-wallet/api/test/test-lib-wallet.js
@@ -3,7 +3,7 @@ import '@agoric/install-ses'; // calls lockdown()
 // eslint-disable-next-line import/no-extraneous-dependencies
 import test from 'tape-promise/tape';
 import bundleSource from '@agoric/bundle-source';
-import { makeIssuerKit, makeLocalAmountMath } from '@agoric/ertp/src/issuer';
+import { makeIssuerKit, makeLocalAmountMath } from '@agoric/ertp';
 
 import { makeZoe } from '@agoric/zoe';
 import fakeVatAdmin from '@agoric/zoe/test/unitTests/contracts/fakeVatAdmin';
