+++ packages/swingset-runner/demo/exchangeBenchmark/bootstrap.js
@@ -1,6 +1,6 @@
 /* global harden */
 
-import makeIssuerKit from '@agoric/ertp';
+import { makeIssuerKit } from '@agoric/ertp';
 import { E } from '@agoric/eventual-send';
 import { makePrintLog } from './printLog';
