+++ packages/swingset-runner/demo/zoeTests/bootstrap.js
@@ -1,7 +1,7 @@
 /* global harden */
 
 import { E } from '@agoric/eventual-send';
-import makeIssuerKit from '@agoric/ertp';
+import { makeIssuerKit } from '@agoric/ertp';
 import buildManualTimer from './manualTimer';
 import { makePrintLog } from './printLog';
