+++ packages/zoe/test/swingsetTests/zoe/bootstrap.js
@@ -1,5 +1,5 @@
 import { E } from '@agoric/eventual-send';
-import makeIssuerKit from '@agoric/ertp';
+import { makeIssuerKit } from '@agoric/ertp';
 import buildManualTimer from '../../../tools/manualTimer';
 
 /* eslint-disable import/no-unresolved, import/extensions */
